requirements.txt
flask
requests
bcrypt
Flask-Bcrypt
Flask-PyMongo
pymongo
python-dotenv